Refactoring Apps Resources & Pages

+++ src/app/models/resource/ResourceFactory.php
@@ -1,6 +1,12 @@
 <?php
 
 require_once DIR . '/app/models/resource/managers/ResourceManager.php';
+require_once DIR . '/app/models/resource/Resource.php';
+require_once DIR . '/app/models/resource/File.php';
+require_once DIR . '/app/models/resource/Folder.php';
+require_once DIR . '/app/models/resource/UnknownResource.php';
+require_once DIR . '/app/models/resource/PlateformeFolder.php';
+require_once DIR . '/app/models/resource/BaseFolder.php';
 
 class ResourceFactory
 {
@@ -26,6 +32,8 @@ class ResourceFactory
 		if (static::$ResourceManager->isDirectory($path)) {
 			if (static::$ResourceManager->isBaseFolder($path)) {
 				return new BaseFolder();
+			} else if (static::$ResourceManager->isPlateformeFolder($path)) {
+				return new PlateformeFolder($path);
 			} else {
 				return new Folder($path);
 			}
@@ -39,8 +47,8 @@ class ResourceFactory
 		if (is_array($path)) {
 			return rtrim(DIR_DATAS . join('/', $path), '/');
 		}
-		if ( strpos($path,DIR_DATAS) === false ){
-			$path = DIR_DATAS.$path;
+		if (strpos($path, DIR_DATAS) === false) {
+			$path = DIR_DATAS . $path;
 		}
 		return $path;
 	}
@@ -48,11 +56,21 @@ class ResourceFactory
 	public static function deleteFromPath($path)
 	{
 		$Resource = static::fromPath($path);
-		static::$ResourceManager->delete($Resource);
+		if (!($Resource instanceof PlateformeFolder) && !($Resource instanceof UnknownResource)) {
+			static::$ResourceManager->delete($Resource);
+		}
 	}
 
 	public static function findAll()
 	{
 		return static::$ResourceManager->findAll();
+	}
+
+	public static function uploadFile($path, $file)
+	{
+		if (isset($file)) {
+			return static::$ResourceManager->addFile(basename($file['name']), $file['tmp_name'], static::cleanPath($path));
+		}
+		return false;
 	}
 }
